+++ src/xulbux/__init__.py
@@ -1,11 +1,21 @@
-__version__ = "1.9.3"
+__package_name__ = "xulbux"
+__version__ = "1.9.4"
+__description__ = "A Python library to simplify common programming tasks."
+__status__ = "Production/Stable"
+
+__url__ = "https://github.com/XulbuX/PythonLibraryXulbuX"
 
 __author__ = "XulbuX"
 __email__ = "xulbux.real@gmail.com"
 __license__ = "MIT"
 __copyright__ = "Copyright (c) 2024 XulbuX"
-__url__ = "https://github.com/XulbuX/PythonLibraryXulbuX"
-__description__ = "A Python library to simplify common programming tasks."
+
+__requires_python__ = ">=3.10.0"
+__dependencies__ = [
+    "keyboard>=0.13.5",
+    "prompt_toolkit>=3.0.41",
+    "regex>=2023.10.3",
+]
 
 __all__ = [
     "Code",
@@ -14,9 +24,9 @@ __all__ = [
     "Data",
     "EnvPath",
     "File",
+    "FileSys",
     "FormatCodes",
     "Json",
-    "Path",
     "Regex",
     "String",
     "System",
@@ -28,9 +38,9 @@ from .console import Console
 from .data import Data
 from .env_path import EnvPath
 from .file import File
+from .file_sys import FileSys
 from .format_codes import FormatCodes
 from .json import Json
-from .path import Path
 from .regex import Regex
 from .string import String
 from .system import System